some better organization: - PlotType is not a global concept, but something only the MPL plot uses - PlotItem.style is replaced with type PlotDataType to streamline - make no assumptions about plotting, leave all details to the backend classes
